git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/smoothvideo@20070 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/settings/VideoSettings.h
@@ -74,7 +74,7 @@ public:
   bool m_bForceIndex;
   EINTERLACEMETHOD m_InterlaceMethod;
   ESCALINGMETHOD   m_ScalingMethod;
-  int m_FilmGrain;
+  float m_FilmGrain;
   int m_ViewMode;   // current view mode
   float m_CustomZoomAmount; // custom setting zoom amount
   float m_CustomPixelRatio; // custom setting pixel ratio
@@ -85,9 +85,9 @@ public:
   float m_SubtitleDelay;
   bool m_SubtitleOn;
   bool m_SubtitleCached;
-  int m_Brightness;
-  int m_Contrast;
-  int m_Gamma;
+  float m_Brightness;
+  float m_Contrast;
+  float m_Gamma;
   float m_NoiseReduction;
   float m_Sharpness;
   float m_AudioDelay;
